Fix/reported issues

+++ lib/features/profile/overview/profiles_overview_notifier.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 import 'package:fpdart/fpdart.dart';
 import 'package:hiddify/core/haptic/haptic_service.dart';
-import 'package:hiddify/core/model/failures.dart';
+import 'package:hiddify/features/connection/notifier/connection_notifier.dart';
 import 'package:hiddify/features/profile/data/profile_data_providers.dart';
 import 'package:hiddify/features/profile/data/profile_repository.dart';
 import 'package:hiddify/features/profile/model/profile_entity.dart';
@@ -47,6 +47,16 @@ class ProfilesOverviewNotifier extends _$ProfilesOverviewNotifier with AppLogger
   Future<void> deleteProfile(ProfileEntity profile) async {
     loggy.debug('deleting profile: ${profile.name}');
 
+    if (profile.active) {
+      final connectionStatus = await ref.read(connectionNotifierProvider.future);
+      if (!connectionStatus.isDisconnected) await ref.read(connectionNotifierProvider.notifier).toggleConnection();
+      final profiles = state.value;
+      if (profiles == null) return;
+      if (profiles.length > 1) {
+        final id = profiles.where((prof) => !prof.active).reduce((a, b) => a.lastUpdate.isAfter(b.lastUpdate) ? a : b).id;
+        await selectActiveProfile(id);
+      }
+    }
     await _profilesRepo.deleteById(profile.id).match(
       (err) {
         loggy.warning('failed to delete profile', err);
